+++ server/api/index.js
@@ -3,7 +3,9 @@ module.exports = router
 
 router.use('/users', require('./users'))
 
-router.use('/orders', require('./orders'))
+router.use('/orders/history', require('./orderHistory'))
+
+router.use('orders/cart', require('./cart'))
 
 router.use((req, res, next) => {
   const error = new Error('Not Found')
