+++ server/api/orderHistory.js
@@ -1,0 +1,3 @@
+const router = require('express').Router()
+const {Order} = require('../db/models')
+module.exports = router
